requirements.txt
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
EasyProcess==1.1
entrypoint2==1.1
executing==1.2.0
filelock==3.9.0
fonttools==4.39.4
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
ipykernel==6.23.1
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
MouseInfo==0.1.3
mpmath==1.2.1
mss==9.0.1
nest-asyncio==1.5.6
networkx==3.0
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
prompt-toolkit==3.0.38
psutil==5.9.5
pure-eval==0.2.2
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
Pygments==2.15.1
PyMsgBox==1.0.9
pynput==1.7.6
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
python-dateutil==2.8.2
pytube==15.0.0
pytweening==1.0.7
pytz==2023.3
pywin32==306
PyYAML==6.0
pyzmq==25.1.0
requests==2.28.1
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
sympy==1.11.1
torch==2.0.1+cu117
torchaudio==2.0.2+cu117
torchvision==0.15.2+cu117
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.4.0
tzdata==2023.3
ultralytics==8.0.115
urllib3==1.26.13
wcwidth==0.2.6
transformers==4.30.1